backend/requirement.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
pydantic
python-dotenv
requests
redis
apscheduler
sentry-sdk
supabase
email-validator
pydantic[email]